requirements.txt
alembic==1.7.6
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==7.1.2
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
importlib-metadata==4.11.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
Werkzeug==2.0.3
zipp==3.7.0
